move some methods to the base class
